refactored location of created components into components folder. renamed theme.js to index in its own folder.
